refactor: remove controller files copy
